% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AudioMeta.R
\name{AudioMeta}
\alias{AudioMeta}
\title{Metadata}
\usage{
AudioMeta(
  Location = NULL,
  Lat = NULL,
  Lon = NULL,
  Device = c("AudioMoth", "Olympus LS-3", "Olympus LS-11"),
  Micro = c("none", "Primo EM-172", "Dodotronics HiSound"),
  Min_conf = 0.7,
  Overlap = 0.25,
  Sensitivity = 1.25,
  Slist = "E:/BirdNET/species_list.txt"
)
}
\arguments{
\item{Location}{location name}

\item{Lat}{latitude}

\item{Lon}{longitude}

\item{Device}{recording device}

\item{Micro}{external microphone}

\item{Min_conf}{min_conf}

\item{Overlap}{overlap}

\item{Sensitivity}{sensitivity}

\item{Slist}{slist}
}
\description{
Add metadata to recording summary. Including
(1) Recording location and parameters
(2) BirdNET analysis settings
}
