% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET.R
\name{BirdNET}
\alias{BirdNET}
\title{BirdNET batch processing}
\usage{
BirdNET(path = NULL, recursive = FALSE, meta = NULL)
}
\arguments{
\item{path}{path to text files}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{meta}{optional. data.frame with recording metadata}
}
\description{
BirdNET batch processing
}
