% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_extract.R
\name{BirdNET_extract}
\alias{BirdNET_extract}
\title{Extract specific event from BirdNET results file}
\usage{
BirdNET_extract(
  path = NULL,
  taxon = NULL,
  score = NULL,
  nmax = NULL,
  sec = 1,
  output = NULL,
  hyperlink = F
)
}
\arguments{
\item{path}{path}

\item{taxon}{optional. select target taxa}

\item{score}{minimum confidence score}

\item{nmax}{maximum number of segments per taxon. default 10}

\item{sec}{seconds before and after target to extract}

\item{output}{path to write wav files}

\item{hyperlink}{optional. Insert hyperlink to audio file in xlsx document. Only if records are not filtered.}
}
\description{
Extract specific event from BirdNET results file
}
\details{
Prerequisites:
(1) Run analyzer.py (same folder specified for --i and --o!)
(2) Reshape output using function BirdNET (this package)
(3) Ensure wav files, BirdNET_results.txt and BirdNET.xlsx files share the same path
}
