% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_results.R
\name{BirdNET_results2txt}
\alias{BirdNET_results2txt}
\title{Modify 'BirdNET.results.txt' created by BirdNET-Analyzer}
\usage{
BirdNET_results2txt(path = NULL)
}
\arguments{
\item{path}{path to text files}
}
\value{
data frame
}
\description{
Tweaks audacity labels by composing a string consisting of species name, recording date and time and detection score (e.g. "Waldwasserläufer 2023-08-15 [0.89]"). Overwrites original text files!
}
\details{
#' Reads text file containing audacity labels that were created by running the BirdNET analyzer script. Note, specifying rtype 'audacity' is required. Output is reformatted by adding the correct time stamp (estimated from file names) to the detected events
}
