% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_species.list.R
\name{BirdNET_species.list}
\alias{BirdNET_species.list}
\title{Create custom species list}
\usage{
BirdNET_species.list(names = NULL, out = NULL, scientific = FALSE)
}
\arguments{
\item{names}{character vector. either scientific names or German species names}

\item{out}{path}

\item{scientific}{logical}
}
\description{
Create custom species list
}
