% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_table.R
\name{BirdNET_table}
\alias{BirdNET_table}
\title{Tabulise BirdNET detections}
\usage{
BirdNET_table(path = NULL, recursive = FALSE)
}
\arguments{
\item{path}{path}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{plot}{ggplot}
}
\description{
Tabulise BirdNET detections
}
\keyword{internal}
