% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_batch_process.R
\name{cleanup_batch_process}
\alias{cleanup_batch_process}
\title{Remove output created while processing audio files}
\usage{
cleanup_batch_process(path = NULL)
}
\arguments{
\item{path}{path}
}
\value{
none.
}
\description{
Removes folder 'split' and all files with extension 'extracted' within the main folder. Wont touch original files and only works in interactive mode by requesting user input
}
