% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_R_script2.R
\name{create_R_script2}
\alias{create_R_script2}
\title{creates 'R_code.R' as a template within given path}
\usage{
create_R_script2(path, open = TRUE)
}
\arguments{
\item{path}{path used for analysing recordings}

\item{open}{by default opens the R Script}
}
\value{
Writes and open script file
}
\description{
creates 'R_code.R' as a template within given path
}
