% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dusk2dawn.R
\name{dusk2dawn}
\alias{dusk2dawn}
\title{Compose dusk-to-dawn string based on date and location}
\usage{
dusk2dawn(date = Sys.Date(), lat = 52.032, lon = 8.517, tz = "CET")
}
\arguments{
\item{date}{Date. Default Sys.Date()}

\item{lat}{latitude}

\item{lon}{longitude}

\item{tz}{timezone. Default UTC}
}
\description{
Compose dusk-to-dawn string based on date and location
}
