% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DateTime.R
\name{get_DateTime}
\alias{get_DateTime}
\title{Get date and time from file name of the form `YYYYMMDD_HHMMSS`}
\usage{
get_DateTime(target, target.path)
}
\arguments{
\item{target}{original recording}

\item{target.path}{original recording path}
}
\value{
data frame
}
\description{
Get date and time from file name of the form `YYYYMMDD_HHMMSS`
}
