% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather.R
\name{get_weather}
\alias{get_weather}
\title{Query weather summary using api.brightsky.dev}
\usage{
get_weather(
  lat = 52.03209,
  lon = 8.516775,
  from = lubridate::as_datetime("2022-06-30 05:00:00"),
  to = lubridate::as_datetime("2022-06-30 10:00:00")
)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{from}{start of query as date time}

\item{to}{end of query as date time}
}
\description{
Query weather summary using api.brightsky.dev
}
\examples{
## test
get_weather()
}
