% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_date_time_name}
\alias{has_date_time_name}
\title{check if file name is composed of date and time string}
\usage{
has_date_time_name(files)
}
\arguments{
\item{files}{character vector}
}
\description{
check if file name is composed of date and time string
}
