% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inspect_events}
\alias{inspect_events}
\title{condense output of find_events}
\usage{
inspect_events(threshold_detection = NULL)
}
\arguments{
\item{threshold_detection}{object of class threshold_detection (see \code{\link[bioacoustics]{threshold_detection}})}
}
\description{
condense output of find_events
}
