% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_events.R
\name{merge_events}
\alias{merge_events}
\title{Merge output}
\usage{
merge_events(path)
}
\arguments{
\item{path}{Path to a set of recordings (all same format and continuous time span).}
}
\description{
Merge output
}
