% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NocMig_meta.R
\name{NocMig_meta}
\alias{NocMig_meta}
\title{Description of dusk, dawn and weather during NocMig recording session}
\usage{
NocMig_meta(lat = 52.03209, lon = 8.516775, date = Sys.Date() - 1)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{date}{start date of NocMig recording}
}
\description{
Description of dusk, dawn and weather during NocMig recording session
}
\details{
(1) Extract dusk and dawn based on date and location
(2) Retrieve weather data and
(3) Compose two string for lists before and past midnight
}
\note{
strings follow the format suggested in https://www.hgon.de/fileadmin/HGONContent/03-Beobachten/07-NocMig/NocMig_Methodenstandardisierung_V1.pdf
}
\references{
Conversion wind direction from degrees to cardinal follows https://community.rstudio.com/t/convert-wind-direction-degrees-into-factors-in-a-data-frame/14636/2

Weather is summarised by computing mean (temp, windspeed) or mode (cloud cover, wind direction) during NocMig recording sessions, seperately for the time before and after midnight.
}
