% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NocMig_process.R
\name{NocMig_process}
\alias{NocMig_process}
\title{Process NocMig session for downstream analysis}
\usage{
NocMig_process(
  path = "E:/NocMig",
  folder = NULL,
  format = c("WAV", "wav", "mp3", "MP3"),
  recorder = c("Olympus LS-3", "Sony PCM-D100", "AudioMoth"),
  time = c("mtime", "ctime"),
  lat = 52.03209,
  lon = 8.516775
)
}
\arguments{
\item{path}{path to NocMig folders}

\item{folder}{folder to process}

\item{format}{Format of sound files (default and suggested is to use WAV).}

\item{recorder}{Currently three templates to ensure correct handling of times. \bold{Only relevant if \code{rename = TRUE}!}.}

\item{time}{Controls, if ctime or mtime is used to compute date_time objects. \bold{Only relevant if \code{rename = TRUE}!}}

\item{lat}{latitude}

\item{lon}{longitude}
}
\description{
Process NocMig session for downstream analysis
}
