% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non_overlapping}
\alias{non_overlapping}
\title{Check for and handle overlapping selections}
\usage{
non_overlapping(df)
}
\arguments{
\item{df}{data frame with event data created by \code{\link[bioacoustics]{threshold_detection}}}
}
\description{
Detects overlapping selections and merges them using \code{\link[warbleR]{overlapping_sels}}
}
