% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecreateDateTime.R
\name{RecreateDateTime}
\alias{RecreateDateTime}
\title{RecreateDateTime from file name}
\usage{
RecreateDateTime(head)
}
\arguments{
\item{head}{character}
}
\description{
RecreateDateTime from file name
}
