% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_recording.R
\name{rename_recording}
\alias{rename_recording}
\title{Rename recording using a string of the form YYYYMMDD_HHMMSS}
\usage{
rename_recording(
  path = NULL,
  recorder = c("Olympus LS-3", "Sony PCM-D100"),
  format = c("WAV", "wav", "mp3", "MP3"),
  time = c("ctime", "mtime"),
  write_text = TRUE,
  simulate = FALSE
)
}
\arguments{
\item{path}{folder containing files (wav or mp3)}

\item{recorder}{Type of recorder}

\item{format}{Format of sound files (default and suggested is to use WAV).}

\item{time}{Either ctime or mtime}

\item{write_text}{logical, if TRUE exports a text file with information about file renaming}

\item{simulate}{logical. If TRUE only shows data frame of names without touching files}
}
\value{
data frame listing the modified file names
}
\description{
For easier manipulation of data obtained using long-term recordings, vendor-specific file names and exchanged by a string of the form `YYYYMMDD_HHMMSS`, as used by the `AudioMoth` (https://www.openacousticdevices.info/). Date and time information is queried using \code{\link[base]{file.info}}. **Note**, popular recorders (e.g., `Olympus LS` or `Sony PCM` series) differ by having the habit of either saving each file one-by-one (i.e., each file has a unique `ctime`) or in batch mode (i.e, all are saved at once).
}
