% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sona_vid.R
\name{sona.vid}
\alias{sona.vid}
\title{Create a dynamic sonagram as video clip using dynaSpec}
\usage{
sona.vid(
  wave = NULL,
  res = 120,
  wl = 1024,
  ymin = 0,
  ymax = 9,
  loop = 1,
  scalecexlab = 0.5,
  scalefontlab = 0.5,
  cexlab = 0.5,
  cexaxis = 0.5,
  t.range = 10,
  tlab = "",
  flab = "kHz",
  colbg = c("black", "white"),
  fix.time = F,
  out = "default"
)
}
\arguments{
\item{wave}{path to a .mp3 or .wav audio file.}

\item{res}{res}

\item{wl}{A numeric vector of length 1 specifying the window length of the spectrogram, default 
is NULL. If supplied, 'hop.size' is ignored.}

\item{ymin}{min frequency}

\item{ymax}{max frequency}

\item{loop}{Logical argument to control if the video is formatted to be played in a loop (i.e. if ends at the start of the clip).}

\item{scalecexlab}{scalecexlab}

\item{scalefontlab}{scalefontlab}

\item{cexlab}{cexlab}

\item{cexaxis}{cexaxis}

\item{t.range}{x-axis width in seconds}

\item{tlab}{name of time (x) axis}

\item{flab}{name of frequency (y) axis}

\item{colbg}{Character string to control the background color. Default is 'white'.}

\item{fix.time}{Logical argument to control if the time axis moves along with the spectrogram or remains fixed. Default is \code{TRUE} (fixed).}

\item{out}{output file name}
}
\description{
Implements dynamic sonagrams as .mp4 files as explained in much detail here (here)[https://marce10.github.io/dynaSpec/] Only works if (FFMPEG)[https://ffmpeg.org/download.html] is installed (see (wikihow tutorial)[https://www.wikihow.com/Install-FFmpeg-on-Windows] to get help)
}
\references{
https://marce10.github.io/dynaSpec/
}
