% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_wave.R
\name{split_wave}
\alias{split_wave}
\title{Simple function to split large audio into segments}
\usage{
split_wave(
  path = NULL,
  file = NULL,
  segment = 600,
  downsample = NULL,
  format = c("WAV", "wav"),
  mono = TRUE
)
}
\arguments{
\item{path}{path}

\item{file}{file name}

\item{segment}{segment length in seconds. Default 600}

\item{downsample}{optional}

\item{format}{file extension (currently only WAV is supported)}

\item{mono}{logical. By default coerces to mono}
}
\value{
none
}
\description{
Cuts audio file into segments for processing and applies resampling and mono conversion if needed.
}
