% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{starting_time2seconds}
\alias{starting_time2seconds}
\title{convert start_time to seconds}
\usage{
starting_time2seconds(t)
}
\arguments{
\item{t}{character vector}
}
\description{
convert start_time to seconds
}
\keyword{internal}
