% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_presets.R
\name{td_presets}
\alias{td_presets}
\title{target-specific templates for use of automatised signal detection}
\usage{
td_presets(
  target = c("Bubo bubo", "Strix aluco", "Glaucidium passerinum", "NocMig")
)
}
\arguments{
\item{target}{name of specified target}
}
\value{
data frame with specified parameters values for use with \code{\link{find_events}} or \code{\link[bioacoustics]{threshold_detection}}
}
\description{
Bundles a few target-specific settings for using \code{\link[bioacoustics]{threshold_detection}} to detect audio signals (*not embarking on identifications*) based on approximate frequency bandwidth and duration of calls.
}
\examples{
## Eagle owl Bubo bubo
td_presets("Bubo bubo")

## Tawny owl Strix aluco
td_presets("Strix aluco")

## Pygmy owl Glaucidium passerinum
td_presets("Glaucidium passerinum")

## NocMig setting
td_presets("NocMig")


}
