% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templt.R
\name{templt}
\alias{templt}
\title{Make template}
\source{
https://marce10.github.io/2020/06/15/Automatic_signal_detection-_a_case_study.html
}
\usage{
templt(wave = NULL, channel = 1, start = NULL, end = NULL)
}
\arguments{
\item{wave}{sound file}

\item{channel}{channel. Default 1}

\item{start}{start in seconds}

\item{end}{stop in seconds}
}
\description{
Make template
}
