% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_events}
\alias{update_events}
\title{Update events: Create threshold_detection from audacity labels}
\usage{
update_events(txt = NULL)
}
\arguments{
\item{txt}{audacity labels}
}
\description{
Update events: Create threshold_detection from audacity labels
}
