% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirdNET_species.list.R
\name{BirdNET_species.list}
\alias{BirdNET_species.list}
\title{Create custom species list}
\usage{
BirdNET_species.list(names = NULL, out = NULL)
}
\arguments{
\item{names}{character vector}

\item{out}{path}
}
\description{
Create custom species list
}
