% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_events.R
\name{extract_events}
\alias{extract_events}
\title{Extract detected events and writes them to a sound file}
\usage{
extract_events(
  threshold_detection,
  buffer = 1,
  format = c("WAV", "wav", "mp3", "MP3"),
  path,
  LPF = NULL,
  HPF = NULL
)
}
\arguments{
\item{threshold_detection}{either class threshold_detection or path to audacity marks}

\item{buffer}{Buffer in seconds added to before and after the event (default 1). Controls also the detection of overlapping events.}

\item{format}{Format of sound files (default WAV).}

\item{path}{where to look up the sound file}

\item{LPF}{integer. Low-Pass Filter (Hz). Frequencies above the cutoff are greatly attenuated.
Default is set internally at the Nyquist frequency of the recording.}

\item{HPF}{integer. High-Pass Filter (Hz). Frequencies below the cutoff are greatly attenuated.
Default setting is 16000 Hz. A default of 1000 Hz is recommended for most bird vocalizations.}
}
\value{
data frame
}
\description{
Extract detected events and writes them to a sound file
}
