% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_events.R
\name{merge_events}
\alias{merge_events}
\title{Merge output}
\usage{
merge_events(path)
}
\arguments{
\item{path}{Path to a set of recordings (all same format and continuous time span). Important note: File are expected to be named using a YYYYMMDD_HHMMSS string or set \code{reaname = TRUE} to allow renaming. Files including the extensions "_extracted.WAV" or "merged_events.WAV" are reserved to write output files and ignored as inputs.}
}
\description{
Merge output
}
